refactor dialog functionality & add AppInfoDialog
