+++ src/hooks/index.ts
@@ -2,4 +2,4 @@ export { default as useAsyncDispatchOnMount } from './useAsyncDispatchOnMount';
 export { default as useSceneKeyboardControls } from './useSceneKeyboardControls';
 export { default as useModelSelectionExport } from './useModelSelectionExport';
 export { default as useHeldRepetitionTimer } from './useHeldRepetitionTimer';
-export { default as useDialogState } from './useDialogState';
+export { default as useDialog } from './useDialog';
